requirements.txt
numpy==1.24.3
pandas==2.0.1
tqdm==4.65.0
matplotlib==3.7.1
scipy==1.10.1
pyyaml==6.0
tensorboard==2.13.0
onnxruntime-gpu
onnx
onnxconverter-common
onnxoptimizer
transformers